Slo sammen brevmottaker typer i en egen fil
