--- src/frontend/Komponenter/Behandling/Brevmottakere/brevmottakerOrganisasjon.ts
@@ -1,5 +1,0 @@
-export interface BrevmottakerOrganisasjon {
-    organisasjonsnummer: string;
-    organisasjonsnavn: string;
-    navnHosOrganisasjon: string;
-}
