--- src/frontend/Komponenter/Behandling/Brevmottakere/brevmottakerPerson.ts
@@ -1,6 +1,0 @@
-import { MottakerRolle } from './mottakerRolle';
-
-export interface BrevmottakerPerson {
-    mottakerRolle: MottakerRolle;
-    navn: string;
-}
