+++ src/frontend/Komponenter/Behandling/Brevmottakere/mottakerRolle.ts
@@ -1,4 +1,3 @@
-import { BrevmottakerPersonUtenIdent } from './baks/brevmottakerPersonUtenIdent';
 import { EøsLandkode } from '../../../Felles/Landvelger/landkode';
 
 export enum MottakerRolle {
@@ -16,33 +15,6 @@ export const mottakerRolleVisningsnavn: Record<MottakerRolle, string> = {
     BRUKER_MED_UTENLANDSK_ADRESSE: 'Bruker med utenlandsk adresse',
     DØDSBO: 'Dødsbo',
 };
-
-export function utledGyldigeMottakerRolle(
-    brevmottakere: BrevmottakerPersonUtenIdent[]
-): MottakerRolle[] {
-    const valgteBrevmottakerRoller = brevmottakere.map(
-        (brevmottaker) => brevmottaker.mottakerRolle
-    );
-
-    if (valgteBrevmottakerRoller.includes(MottakerRolle.DØDSBO)) {
-        return [];
-    }
-
-    if (valgteBrevmottakerRoller.includes(MottakerRolle.BRUKER_MED_UTENLANDSK_ADRESSE)) {
-        return [MottakerRolle.VERGE, MottakerRolle.FULLMAKT];
-    }
-
-    if (
-        valgteBrevmottakerRoller.length > 0 &&
-        !valgteBrevmottakerRoller.includes(MottakerRolle.BRUKER_MED_UTENLANDSK_ADRESSE)
-    ) {
-        return [MottakerRolle.BRUKER_MED_UTENLANDSK_ADRESSE];
-    }
-
-    return Object.values(MottakerRolle)
-        .filter((mottakerRolle) => mottakerRolle !== MottakerRolle.BRUKER)
-        .filter((mottakerRolle) => !valgteBrevmottakerRoller.includes(mottakerRolle));
-}
 
 export function skalNavnVærePreutfyltForMottakerRolle(mottakerRolle: MottakerRolle): boolean {
     return (
